Import Boilerplate Database(s)

This function imports one or more boilerplate databases from disk.

Usage

boilerplate_import(category = NULL, data_path = NULL, quiet = FALSE)

Arguments

category

Character or character vector. Category of database to import. Options include "measures", "methods", "results", "discussion", "appendix", "template". If NULL (default), imports all available categories.

data_path

Character. Base path for data directory. If NULL (default), uses here::here("boilerplate", "data").

quiet

Logical. If TRUE, suppresses all CLI alerts. Default is FALSE.

Value

List. The imported database(s). If a single category was requested, returns that database. If multiple categories were requested, returns a named list with each category's database.

Examples

# import just the methods database
methods_db <- boilerplate_import("methods")
#> ℹ using default path: /Users/joseph/GIT/boilerplate/boilerplate/data
#> ! data directory does not exist: /Users/joseph/GIT/boilerplate/boilerplate/data
#> ℹ importing methods database
#> ! methods database file not found, using default

# import multiple specific databases
dbs <- boilerplate_import(c("methods", "measures"))
#> ℹ using default path: /Users/joseph/GIT/boilerplate/boilerplate/data
#> ! data directory does not exist: /Users/joseph/GIT/boilerplate/boilerplate/data
#> ℹ importing methods database
#> ! methods database file not found, using default
#> ℹ importing measures database
#> ! measures database file not found, using default
methods_db <- dbs$methods
measures_db <- dbs$measures

# import all databases
all_dbs <- boilerplate_import()
#> ℹ importing all categories
#> ℹ using default path: /Users/joseph/GIT/boilerplate/boilerplate/data
#> ! data directory does not exist: /Users/joseph/GIT/boilerplate/boilerplate/data
#> ℹ importing measures database
#> ! measures database file not found, using default
#> ℹ importing methods database
#> ! methods database file not found, using default
#> ℹ importing results database
#> ! results database file not found, using default
#> ℹ importing discussion database
#> ! discussion database file not found, using default
#> ℹ importing appendix database
#> ! appendix database file not found, using default
#> ℹ importing template database
#> ! template database file not found, using default